Use native register width integer in the core loop
I noticed that there was a `uxth` (sign extending) instruction
in the hot loop of the code, because the native registers
are 32 bits but we were using a 16-bit type for the counter.
In this case, gcc emits two instructions (inc + uxth) to
increment `pulse_count` when just one would suffice.

At 120MHz, this changes the peaks from ~21/22, 34/35, 48
to 25/26, 41/42, 57.