greaseweazel: improve index pulse recording
ripping with genuine GW hardware, a typical track is read in
to gw as
```
Flux: 72.00 MHz
 Total: 232202 samples, 594.17ms
 Revolution 0: 93.06ms
 Revolution 1: 166.87ms
 Revolution 2: 166.87ms
 Revolution 3: 166.87ms
```

our firmware came in as
```
Flux: 24.00 MHz
 Total: 120485 samples, 652.35ms
 Revolution 0: 0.00ms
 Revolution 1: 162.34ms
 Revolution 2: 55.11ms
 Revolution 3: 162.17ms
 Revolution 4: 55.28ms
 Revolution 5: 162.12ms
```
Notice how there's an empty revolution 0 and a partial revolution
2 & 4.

It appears that
 * if capture_ms is 0, nothing is captured after the index
   (rather than 50ms)
 * no index is emitted at the start (though this is because
   the first revolution is almost always a partial revolution)

Now after this change, a track on rp2040 looks like
```
Flux: 24.00 MHz
 Total: 96498 samples, 502.46ms
 Revolution 0: 0.00ms
 Revolution 1: 167.49ms
 Revolution 2: 167.49ms
 Revolution 3: 167.49ms
```
which appears to be closer to what gw expects; it also fixes how
g64conv used to require `r1` to select a particular revolution.
This may have been because it defaulted to one of the 'runt'
revolutions, which also made it misestimate the rotational speed
of the drive.