Fix flux reading when index pulse doesn't come
It's possible for `capture_track` to return early, in the
case that the buffer is small relative to the track time
(or something else is going wrong, like no index pulse).

However, in this case the flux sending loop would never exit,
but would index beyond the end of the flux_transitions array
and probably HardFault eventually.

Instead, use (-1) as the sentinel value for 'no index pulse arrived'
and break out of the loop.