Drop the QUARTERTRACK type, APPLE2 is always quarter-track
GW and FluxEngine folks agreed that it is better to just always use
the drive's native step, and FluxEngine has started fixing things
internally so that this works better than when I first implemented the
"why not both" approach.  Only the "3" bustype is allocated now in the
GW protocol.

We can decide if we want to keep the step_mode at all, since its only
intended user was the GW-compatible firmware.