mfm_impl: A full re-write, adds encoding
The old implementation literally decoded MFM on the fly which was cool
but irrelevant: we can store a track worth of flux on all the micros
we care about, in addition to the decoded sector data.

The new implementation also includes an encoder, which writes data
that is understood by a USB 3.5" floppy drive. The MSD demo is enhanced
for read/write support.